Move test JSON (de)serializer to its own file
