+++ Tests/GRPCNIOTransportHTTP2Tests/ControlMessages.swift
@@ -133,22 +133,3 @@ enum CancellationKind: Codable {
 
 struct Empty: Codable {
 }
-
-struct JSONSerializer<Message: Encodable>: MessageSerializer {
-  private let encoder = JSONEncoder()
-
-  func serialize<Bytes: GRPCContiguousBytes>(_ message: Message) throws -> Bytes {
-    let data = try self.encoder.encode(message)
-    return Bytes(data)
-  }
-}
-
-struct JSONDeserializer<Message: Decodable>: MessageDeserializer {
-  private let decoder = JSONDecoder()
-
-  func deserialize<Bytes: GRPCContiguousBytes>(_ serializedMessageBytes: Bytes) throws -> Message {
-    try serializedMessageBytes.withUnsafeBytes {
-      try self.decoder.decode(Message.self, from: Data($0))
-    }
-  }
-}
